# evidence/ex13_capstone/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API de Agendamento Clínico",
    "description": "Pacientes, profissionais de saúde e consultas. Dados de saúde (LGPD art. 11).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/patient/new": {
      "post": {
        "tags": [
          "Patients"
        ],
        "summary": "Create Patient",
        "operationId": "create_patient_patient_new_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientPublic"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "404": {
            "description": "Recurso inexistente OU pertencente a outro usuário (indistinguíveis de propósito)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/patient/": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "List Patients",
        "operationId": "list_patients_patient__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientPublic"
                  },
                  "title": "Response List Patients Patient  Get"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "404": {
            "description": "Recurso inexistente OU pertencente a outro usuário (indistinguíveis de propósito)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patient/search": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Search Patients",
        "operationId": "search_patients_patient_search_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-zÀ-ÖØ-öø-ÿ .'\\-]{2,60}$",
              "description": "Parte do nome (2–60 letras)",
              "title": "Name"
            },
            "description": "Parte do nome (2–60 letras)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientPublic"
                  },
                  "title": "Response Search Patients Patient Search Get"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "404": {
            "description": "Recurso inexistente OU pertencente a outro usuário (indistinguíveis de propósito)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patient/{patient_id}": {
      "get": {
        "tags": [
          "Patients"
        ],
        "summary": "Read Patient",
        "operationId": "read_patient_patient__patient_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientPublic"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "404": {
            "description": "Recurso inexistente OU pertencente a outro usuário (indistinguíveis de propósito)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientCreate": {
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[A-Za-zÀ-ÖØ-öø-ÿ][A-Za-zÀ-ÖØ-öø-ÿ .'\\-]{1,99}$",
            "title": "Name"
          },
          "cpf": {
            "type": "string",
            "maxLength": 14,
            "pattern": "^\\d{3}\\.?\\d{3}\\.?\\d{3}-?\\d{2}$",
            "title": "Cpf"
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+?\\d{10,13}$",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "professional_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Professional Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "cpf",
          "phone",
          "professional_id"
        ],
        "title": "PatientCreate"
      },
      "PatientPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cpf": {
            "type": "string",
            "title": "Cpf"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "professional_id": {
            "type": "integer",
            "title": "Professional Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "cpf",
          "phone",
          "professional_id"
        ],
        "title": "PatientPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/user/signin"
          }
        }
      },
      "LabClientCredentials": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "scopes": {
              "availability:read": "Consultar horários livres de um profissional"
            },
            "tokenUrl": "/oauth/token"
          }
        }
      }
    }
  }
}